+++ bogo/config.json.in
@@ -10,10 +10,10 @@
     "*-RSA_WITH_AES_128_GCM_SHA256-*": "no rsa kem",
     "*-RSA_WITH_AES_256_CBC_SHA-*": "no rsa kem",
     "*-RSA_WITH_AES_256_GCM_SHA384-*": "no rsa kem",
-    "*-Sign-ML-DSA-*": "TODO: no ML-DSA signing support",
-    "*-Sign-Negotiate-ML-DSA-*": "TODO: no ML-DSA signing support",
+#if defined(RING)
+    "*-ML-DSA-*": "no ML-DSA verification support in *ring*",
+#endif
     "*-Sign-RSA_PKCS1_SHA1-*": "no sha1",
-    "*-SignDefault-ML-DSA-*": "TODO: no ML-DSA signing support",
     "*-SignDefault-RSA_PKCS1_SHA1-*": "no sha1",
     "*-TLS1": "we only support >=TLS1.2",
     "*-TLS11": "we only support >=TLS1.2",
@@ -164,10 +164,16 @@
     "Compliance-fips-202205-TLS-Client-Ed25519": "fips approved in aws-lc-rs",
     "Compliance-fips-202205-TLS-Client-MLKEM": "fips pending in aws-lc-rs",
     "Compliance-fips-202205-TLS-Client-X25519MLKEM768": "ml-kem fips pending in aws-lc-rs",
+    "Compliance-fips-202205-TLS-Client-ML-DSA-44": "ML-DSA postdates fips-202205 policy",
+    "Compliance-fips-202205-TLS-Client-ML-DSA-65": "ML-DSA postdates fips-202205 policy",
+    "Compliance-fips-202205-TLS-Client-ML-DSA-87": "ML-DSA postdates fips-202205 policy",
     "Compliance-fips-202205-TLS-Server-ECDSA_P521_SHA512": "fips approved in aws-lc-rs",
     "Compliance-fips-202205-TLS-Server-Ed25519": "fips approved in aws-lc-rs",
     "Compliance-fips-202205-TLS-Server-MLKEM": "fips pending in aws-lc-rs",
     "Compliance-fips-202205-TLS-Server-X25519MLKEM768": "ml-kem fips pending in aws-lc-rs",
+    "Compliance-fips-202205-TLS-Server-ML-DSA-44": "ML-DSA postdates fips-202205 policy",
+    "Compliance-fips-202205-TLS-Server-ML-DSA-65": "ML-DSA postdates fips-202205 policy",
+    "Compliance-fips-202205-TLS-Server-ML-DSA-87": "ML-DSA postdates fips-202205 policy",
 #endif
 
     "__LAST__": ""
@@ -193,14 +199,23 @@
     "CertificateCipherMismatch-ECDSA": ":WRONG_SIGNATURE_TYPE:",
     "CertificateCipherMismatch-Ed25519": ":WRONG_SIGNATURE_TYPE:",
     "CertificateCipherMismatch-RSA": ":WRONG_SIGNATURE_TYPE:",
+    "Client-Sign-ML-DSA-44-TLS12": ":NO_COMMON_SIGNATURE_ALGORITHMS:",
+    "Client-Sign-ML-DSA-65-TLS12": ":NO_COMMON_SIGNATURE_ALGORITHMS:",
+    "Client-Sign-ML-DSA-87-TLS12": ":NO_COMMON_SIGNATURE_ALGORITHMS:",
     "Client-Sign-RSA_PKCS1_SHA256-TLS13": ":INCOMPATIBLE:",
     "Client-Sign-RSA_PKCS1_SHA384-TLS13": ":INCOMPATIBLE:",
     "Client-Sign-RSA_PKCS1_SHA512-TLS13": ":INCOMPATIBLE:",
+    "Client-SignDefault-ML-DSA-44-TLS12": ":NO_COMMON_SIGNATURE_ALGORITHMS:",
+    "Client-SignDefault-ML-DSA-65-TLS12": ":NO_COMMON_SIGNATURE_ALGORITHMS:",
+    "Client-SignDefault-ML-DSA-87-TLS12": ":NO_COMMON_SIGNATURE_ALGORITHMS:",
     "Client-SignDefault-RSA_PKCS1_SHA256-TLS13": ":INCOMPATIBLE:",
     "Client-SignDefault-RSA_PKCS1_SHA384-TLS13": ":INCOMPATIBLE:",
     "Client-SignDefault-RSA_PKCS1_SHA512-TLS13": ":INCOMPATIBLE:",
     "Client-TLS13-NoSign-RSA_PKCS1_MD5_SHA1": ":INCOMPATIBLE:",
     "Client-TooLongSessionID": ":BAD_HANDSHAKE_MSG:",
+    "Client-Verify-ML-DSA-44-TLS12": ":WRONG_SIGNATURE_TYPE:",
+    "Client-Verify-ML-DSA-65-TLS12": ":WRONG_SIGNATURE_TYPE:",
+    "Client-Verify-ML-DSA-87-TLS12": ":WRONG_SIGNATURE_TYPE:",
     "Client-VerifyDefault-ML-DSA-44-TLS12": ":WRONG_SIGNATURE_TYPE:",
     "Client-VerifyDefault-ML-DSA-65-TLS12": ":WRONG_SIGNATURE_TYPE:",
     "Client-VerifyDefault-ML-DSA-87-TLS12": ":WRONG_SIGNATURE_TYPE:",
@@ -255,9 +270,15 @@
     "SendBogusAlertType": ":BAD_ALERT:",
     "SendInvalidRecordType": ":GARBAGE:",
     "Server-NonEmptyEndOfEarlyData-TLS13": ":BAD_HANDSHAKE_MSG:",
+    "Server-Sign-ML-DSA-44-TLS12": ":INCOMPATIBLE:",
+    "Server-Sign-ML-DSA-65-TLS12": ":INCOMPATIBLE:",
+    "Server-Sign-ML-DSA-87-TLS12": ":INCOMPATIBLE:",
     "Server-Sign-RSA_PKCS1_SHA256-TLS13": ":INCOMPATIBLE:",
     "Server-Sign-RSA_PKCS1_SHA384-TLS13": ":INCOMPATIBLE:",
     "Server-Sign-RSA_PKCS1_SHA512-TLS13": ":INCOMPATIBLE:",
+    "Server-SignDefault-ML-DSA-44-TLS12": ":INCOMPATIBLE:",
+    "Server-SignDefault-ML-DSA-65-TLS12": ":INCOMPATIBLE:",
+    "Server-SignDefault-ML-DSA-87-TLS12": ":INCOMPATIBLE:",
     "Server-SignDefault-RSA_PKCS1_SHA256-TLS13": ":INCOMPATIBLE:",
     "Server-SignDefault-RSA_PKCS1_SHA384-TLS13": ":INCOMPATIBLE:",
     "Server-SignDefault-RSA_PKCS1_SHA512-TLS13": ":INCOMPATIBLE:",
@@ -265,6 +286,9 @@
     "Server-TLS13-NoSign-RSA_PKCS1_MD5_SHA1": ":INCOMPATIBLE:",
     "Server-TooLongSessionID-TLS12": ":BAD_HANDSHAKE_MSG:",
     "Server-TooLongSessionID-TLS13": ":BAD_HANDSHAKE_MSG:",
+    "Server-VerifyDefault-ML-DSA-44-TLS12": ":WRONG_SIGNATURE_TYPE:",
+    "Server-VerifyDefault-ML-DSA-65-TLS12": ":WRONG_SIGNATURE_TYPE:",
+    "Server-VerifyDefault-ML-DSA-87-TLS12": ":WRONG_SIGNATURE_TYPE:",
     "ServerAuth-NoFallback-TLS13": ":INCOMPATIBLE:",
     "ServerCipherFilter-ECDSA": ":INCOMPATIBLE:",
     "ServerCipherFilter-Ed25519": ":INCOMPATIBLE:",
@@ -319,6 +343,12 @@
     "CertificateSelection-Client-SignatureAlgorithmECDSACurve-MatchNone-TLS-TLS13": "",
     "CertificateSelection-Client-SignatureAlgorithmKeyPrefs-MatchNone-TLS-TLS12": "",
     "CertificateSelection-Client-SignatureAlgorithmKeyPrefs-MatchNone-TLS-TLS13": "",
+    "Client-Sign-ML-DSA-44-TLS12": "",
+    "Client-Sign-ML-DSA-65-TLS12": "",
+    "Client-Sign-ML-DSA-87-TLS12": "",
+    "Client-SignDefault-ML-DSA-44-TLS12": "",
+    "Client-SignDefault-ML-DSA-65-TLS12": "",
+    "Client-SignDefault-ML-DSA-87-TLS12": "",
     "D__": "",
     "DuplicateExtensionClient-TLS-TLS12": "remote error: illegal parameter",
     "DuplicateExtensionClient-TLS-TLS13": "remote error: illegal parameter",
